refactor(tool-proxy): extract test module to tests_main.rs
Move 332-line inline test module from main.rs to separate file.
main.rs: 4439 → 4110 lines (329 line reduction).
